+++ app/controllers/wicked_pdf_controller.rb
@@ -1,0 +1,23 @@
+# Demo for wicked pdf
+class WickedPdfController < ApplicationController
+  def table_with_no_break
+    respond_to do |format|
+      format.html
+      format.pdf { render pdf_params 'table_with_no_break' }
+    end
+  end
+
+  private
+
+  def pdf_params(template)
+    {
+      pdf:          template,
+      show_as_html: params.key?('debug'),
+      footer: {
+        left:       'CodeRails',
+        right:      '[page] of [topage]',
+        line:       true
+      }
+    }
+  end
+end
