+++ app/helpers/application_helper.rb
@@ -1,2 +1,5 @@
 module ApplicationHelper
+  def index_action?
+    action_name == 'index'
+  end
 end
