+++ app/views/fusion_charts/column2d.html.haml
