+++ app/views/pages/_categories.html.haml
@@ -1,6 +1,6 @@
 = render 'shared/list', title: 'PDF Generation',
          examples: [{ title: 'Wicked PDF',
-                      path: table_with_no_break_demo_wicked_pdf_path}]
+                      path: table_with_no_break_wicked_pdf_path}]
 = render 'shared/list', title: 'Charts',
          examples: [{ title: 'Fusion Charts',
-                      path: column2d_demo_fusion_charts_path}]
+                      path: column2d_fusion_charts_path}]
