+++ app/views/shared/_breadcrumb.html.haml
@@ -1,9 +1,8 @@
 %ol.breadcrumb
-  - if current_page?(root_path)
-    %li.active Home
-    %li.active Demo
-  - else
-    %li= link_to 'Home', root_path
-    %li= link_to 'Demo', root_path
+  %li= link_to 'Demo', root_path
+  - unless controller_name == PagesController.controller_name
     %li.active
-      = yield :action
+      = link_to_unless index_action?, controller_name.humanize, url_for(action: 'index')
+    - unless index_action?
+      %li.active
+        = action_name.humanize
