+++ app/views/wicked_pdf/_table_content.html.haml
