+++ app/views/wicked_pdf/table_with_no_break.html.haml
@@ -4,7 +4,7 @@
   Wicked pdf
 .pdf-button.pull-right
   = link_to 'PDF Version',
-            table_with_no_break_demo_wicked_pdf_path(format: :pdf),
+            table_with_no_break_wicked_pdf_path(format: :pdf),
             class: 'btn btn-primary ', target: :_blank
   .clearfix
 = render 'table_content'
