+++ app/views/wicked_pdf/table_with_no_break.pdf.haml
