+++ spec/controllers/fusion_charts_controller_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe Demo::FusionChartsController, type: :controller do
+RSpec.describe FusionChartsController, type: :controller do
   describe 'GET #column2d' do
     it 'returns http success' do
       get :column2d
