+++ spec/views/fusion_charts/index.html.haml_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe 'fusion_charts/index.html.haml', type: :view do
+  pending "add some examples to (or delete) #{__FILE__}"
+end
